Switching from h2 to Postgre

+++ src/main/java/com/hw/networkstore/model/BaseEntity.java
@@ -7,18 +7,16 @@ public abstract class BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     public String name;
 
     public int price;
 
-    public Long getId() {
-        return id;
+    public int getId() {return id;
     }
 
-    public void setId(Long id) {
-        this.id = id;
+    public void setId(int id) {this.id = id;
     }
 
     public String getName() {
